Moved HttpClientTest and HttpProxyTest back into singletests after fixing them both to work with changes in the master branch.
